Prevented user admin from deleting accounts with id < 1000

+++ subsystem/useradmin/nav/web/useradmin/views.py
@@ -102,6 +102,10 @@ def account_delete(request, account_id):
         # FIXME add message
         return HttpResponseRedirect(reverse('useradmin-account_list'))
 
+    if account.is_system_account():
+        # FIXME add message
+        return HttpResponseRedirect(reverse('useradmin-account_list'))
+
     if request.method == 'POST':
         account.delete()
         # FIXME add message
